+++ src/engine/worldio.cpp
@@ -691,15 +691,13 @@ bool save_world(const char *mname, bool nolms)
 
             // Now comes the good stuff, our own custom attributes.
             if (tmp.type == GAMEENTITY) {
-                j[i]["classname"] = tmp.classname;
-/*                j[i]["str_attr1"] = std::string(tmp.str_attr1);
-                j[i]["str_attr2"] = std::string(tmp.str_attr2);
-                j[i]["str_attr3"] = std::string(tmp.str_attr3);
-                j[i]["str_attr4"] = std::string(tmp.str_attr4);
-                j[i]["str_attr5"] = std::string(tmp.str_attr5);
-                j[i]["str_attr6"] = std::string(tmp.str_attr6);
-                j[i]["str_attr7"] = std::string(tmp.str_attr7);
-                j[i]["str_attr8"] = std::string(tmp.str_attr8);*/
+                // Store classname.
+                j[i]["game"]["classname"] = tmp.classname;
+
+                // Store attributes.
+                for (std::map<std::string, std::string>::iterator k = tmp.attributes.begin(); k != tmp.attributes.end(); ++k) {
+                    j[i]["game"]["attributes"][k->first] = k->second;
+                }
             }
         }
     }
@@ -859,7 +857,7 @@ bool load_world(const char *mname, const char *cname)        // still supports a
 
             // Have to do this here to ensure that classname can be passed to newgameentity.
             if (type == GAMEENTITY) {
-               classname = element["classname"];
+               classname = element["game"]["classname"];
             }
 
             // Allocate our entity.
@@ -881,15 +879,12 @@ bool load_world(const char *mname, const char *cname)        // still supports a
             if (type == GAMEENTITY) {
                 // Store the classname.
                 e.classname = classname;
+                //e.attributes = element.at("game").at("attributes");
+                const json& rh = element["game"]["attributes"];
 
-/*                copycubestr(&e.str_attr1[0], std::string(element["str_attr1"]).c_str(), 256);
-                copycubestr(&e.str_attr2[0], std::string(element["str_attr2"]).c_str(), 256);
-                copycubestr(&e.str_attr3[0], std::string(element["str_attr3"]).c_str(), 256);
-                copycubestr(&e.str_attr4[0], std::string(element["str_attr4"]).c_str(), 256);
-                copycubestr(&e.str_attr5[0], std::string(element["str_attr5"]).c_str(), 256);
-                copycubestr(&e.str_attr6[0], std::string(element["str_attr6"]).c_str(), 256);
-                copycubestr(&e.str_attr7[0], std::string(element["str_attr7"]).c_str(), 256);
-                copycubestr(&e.str_attr8[0], std::string(element["str_attr8"]).c_str(), 256);*/
+                for (auto& element : json::iterator_wrapper(rh)) {
+                    e.attributes[element.key()] = element.value();
+                }
             }
 
             ents.add(&e);